update Tabs story
