--- storybook/src/Tabs/Controlled.stories.tsx
@@ -1,62 +1,0 @@
-/*
-
- MIT License
-
- Copyright (c) 2020 Looker Data Sciences, Inc.
-
- Permission is hereby granted, free of charge, to any person obtaining a copy
- of this software and associated documentation files (the "Software"), to deal
- in the Software without restriction, including without limitation the rights
- to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the Software is
- furnished to do so, subject to the following conditions:
-
- The above copyright notice and this permission notice shall be included in all
- copies or substantial portions of the Software.
-
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- SOFTWARE.
-
- */
-
-import React, { FC, useState } from 'react'
-import {
-  Tab,
-  Tabs,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Space,
-} from '@looker/components'
-
-export default {
-  title: 'Tabs/Controlled',
-}
-
-export const Controlled: FC = () => {
-  const [currentTabIndex, setTab] = useState(0)
-
-  return (
-    <>
-      <Space>
-        <button onClick={() => setTab(0)}>Go to A</button>
-        <button onClick={() => setTab(1)}>Go to B</button>
-      </Space>
-      <Tabs index={currentTabIndex} onChange={setTab}>
-        <TabList>
-          <Tab>A</Tab>
-          <Tab>B</Tab>
-        </TabList>
-        <TabPanels>
-          <TabPanel>A</TabPanel>
-          <TabPanel>B</TabPanel>
-        </TabPanels>
-      </Tabs>
-    </>
-  )
-}
